Get jvox working pretty much like cvox.
The flushes out a bunch of corner cases around how scope is handled.
Trying to find the right balance between not wanting to do resolution
in jvox, but not wanting fishy scope rules for cvox.

Still a couple of things to hash at around duplicate locals in the same
scope, but it's mostly there.

Also change how fields and methods are handled to match cvox. Now only
instance have fields.