Do some clean up on Environment.
It was using a mixture of get()/set() and getAt()/setAt(), even after
the resolution pass. The get()/set() code which walks enclosing
environments should never be needed after resolution has been performed,
so I cleaned that up.

Still not super crazy about local scopes are expressed, but it's a
little better.