Decide what to do about for loops.
- Not going to do lists and iterables. Too much grunt work for too
  little elightenment.
- Instead, for() is a C-style loop.
- Desugars to a while() loop, which is a good excuse to talk about
  desugaring to core forms.
- Also reworked resolver a bit.

Alas, this means we need access to blocks in the Control Flow chapter,
so this breaks a bunch of stuff. I'm going to have to either move for
out of Control Flow to a later chapter (maybe split looping and
branching?) or push block scope forward before functions.

That would probably mean the "Blocks and Bindings" chapter becomes
"Closures" (again).