Include code sections in Markdown.
The big chunk of work here is adding support for multiple sections for
the same chapter, each with its own number. Before, sections only had
chapter granularity.

I added support for numbering the sections. Then I move the code for
parsing the sections into a separate little module so that it can be
reused for splitting the chapters and for including code snippets.