Rework code to allow named snippets.
Having them be explicitly numbered is too annoying when I want to
insert a snippet before some that I've already split out. This allows
them to have arbitrary names.

We still need to what order the snippets appear, so this parses the
".md" to find all of the "^code" tags and orders based on how they
appear there.

There's still a bunch of clean up and testing to do, but it looks like
this is producing the same output as the old code.